feat: add authorizes for category + remove baseController

+++ app/Http/Controllers/Api/CategoryController.php
@@ -1,15 +1,17 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use App\Http\Resources\CategoryCollection;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
-use App\Http\Controllers\BaseController;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
+use Gate;
+use Illuminate\Auth\Access\AuthorizationException;
 
-class CategoryController extends BaseController
+class CategoryController extends Controller
 {
     /**
      * Display a listing of the resource.
@@ -28,6 +30,10 @@ class CategoryController extends BaseController
      */
     public function store(StoreCategoryRequest $request)
     {
+        if (Gate::denies('role.manager')) {
+            return $this->sendUnauthorized();
+        }
+
         $data = $request->validated();
         $result = Category::create($data);
         return $this->sendResponse("Category Created", 200, new CategoryResource($result));
@@ -46,6 +52,10 @@ class CategoryController extends BaseController
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+        if (Gate::denies('role.manager')) {
+            return $this->sendUnauthorized();
+        }
+
         $data = $request->validated();
         $result = $category::update($data);
         return $this->sendResponse("Category Updated", 200, new CategoryResource($result));
@@ -56,6 +66,10 @@ class CategoryController extends BaseController
      */
     public function destroy(Category $category)
     {
+        if (Gate::denies('role.manager')) {
+            return $this->sendUnauthorized();
+        }
+
         $category->delete();
         return $this->sendResponse("Category Deleted", 200);
     }
